+++ src/Streamr.ts
@@ -1,6 +1,6 @@
 import StreamrClient from 'streamr-client'
-import detectEthereumProvider from '@metamask/detect-provider'
-import { STREAMR_SESSION_TOKEN } from './Config'
+import createMetaMaskProvider  from 'metamask-extension-provider'
+import { STREAMR_SESSION_TOKEN, STREAMR_API_ENDPOINT } from './Config'
 
 import type { AbstractProvider } from 'web3-core'
 
@@ -9,27 +9,22 @@ interface Auth {
     provider?: object
 }
 
-interface Filter {
-    last?: number,
-    from?: { timestamp: number },
-    to?: { timestamp: number }
-}
-
 interface Permission {
     id: number,
     user: string,
     operation: string
 }
 
-export const streamrCollect = async (streamId: string, filter: Filter): Promise<object[]> => {
+export const streamrCollect = async (streamId: string, count: number): Promise<object[]> => {
     return new Promise( async (resolve, reject) => {
-        console.log('Create stream snapshot ...')
+        window['streamr2ocean'].status = 'Retrieve stream snapshot ...'
 
-        const provider = <AbstractProvider> await detectEthereumProvider()
+        const provider = <AbstractProvider> await createMetaMaskProvider()
+
+        provider['autoRefreshOnNetworkChange'] = false
 
         await provider.request!({ method: 'eth_requestAccounts' })
 
-        let data = <object[]> []
         let sessionToken = <string> getSessionToken()
 
         const auth = <Auth> sessionToken ? { sessionToken } : { provider }
@@ -43,24 +38,17 @@ export const streamrCollect = async (streamId: string, filter: Filter): Promise<
         const stream = await client.getStream(streamId)
 
         if (!stream) {
-            return reject(new Error('The stream you were looking for was not found.'))
+            return reject(new Error('The stream was not found'))
         }
 
         if (! await hasSharePermission(stream)) {
-            return reject(new Error('The stream does not have share permission.'))
+            return reject(new Error('Share permission missing'))
         }
 
-        const subscription = client.subscribe(
-            { stream: streamId, resend: filter },
-            (message: object) => data.push(message)
-        )
-
-        subscription.on('resent', () => client.disconnect())
-        subscription.on('no_resend', () => client.disconnect())
-        subscription.on('error', error => reject(error))
-
-        client.on('disconnected', () => resolve(data))
-        client.on('error', error => reject(error))
+        // No client.subscribe(), because Chrome Extensions can't handle WebSockets
+        fetchDataWithoutWebsocket(streamId, count)
+            .then(data => resolve(data))
+            .catch(error => reject(error))
     })
 }
 
@@ -76,4 +64,38 @@ const hasSharePermission = async (stream): Promise<boolean> => {
     const permissions = <Permission[]> await stream.getMyPermissions()
 
     return !!permissions.filter(e => e.operation === 'stream_share').length
+}
+
+const fetchDataWithoutWebsocket = async (streamId: string, count: number): Promise<object[]> => {
+    const apiEndpointUrl = STREAMR_API_ENDPOINT
+        .replace('%streamId%', encodeURIComponent(streamId))
+        .replace('%count%', encodeURIComponent(count))
+
+    return new Promise( async (resolve, reject) => {
+        fetchRetry(apiEndpointUrl, 5)
+            .then(json => resolve(json.map(item => item['content'])))
+            .catch(error => reject(error))
+    })
+}
+
+const fetchRetry = async (url: string, retries: number): Promise<object[]> => {
+    return new Promise( async (resolve, reject) => {
+        fetch(url).then(async response => {
+            if (!response.ok || response.status !== 200) {
+                return reject(new Error('The stream has no content'))
+            }
+
+            const json = await response.json()
+
+            if (json.length) {
+                return resolve(json)
+            }
+
+            if (retries > 0) {
+                console.log('No content received from API, try again', retries)
+
+                return fetchRetry(url, retries - 1)
+            }
+        })
+    })
 }