+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Changelog
 
+#### 0.2.0 (2021-01-02)
+* Add Chromium extension
+* Remove Svelte from project
+* Add publishing process status messages
+* Replace `client.subscribe()` by `fetch()` 
+* Replace `@metamask/detect-provider` by `metamask-extension-provider`
+
 #### 0.1.0 (2020-12-30)
 * Add check for `stream_share` permission